+++ packages/moonstone/CHANGELOG.md
@@ -18,6 +18,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/VirtualList` not to let the focus move outside of container even if there are children left when navigating with 5way
 - `moonstone/Scrollable` to update disability of paging controls when the scrollbar is set to `visible` and the content becomes shorter.
 - `moonstone/VideoPlayer` to focus on hover over play/pause button when video is loading.
+- `moonstone/IncrementSlider` and `moonstone/Picker` to read value changes when selecting buttons
 
 ## [1.5.0] - 2017-07-19